# backend/swagger.yaml
openapi: 3.0.0
info:
  title: QuizMaster Pro API (backend)
  version: "1.0"
  description: API documentation for the QuizMaster Pro application's backend server
paths:
...
  /api/users/{id}:
    put:
      summary: Update user
      description: Update user information
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [user, admin]
      responses:
        "200":
          description: OK
    delete:
      summary: Delete user
      description: Delete a user
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
...